+++ src/app/app.component.ts
@@ -1,10 +1,17 @@
-import { Component } from '@angular/core';
+import { AfterViewInit, Component, ElementRef, ViewChild } from '@angular/core';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.css']
 })
-export class AppComponent {
+export class AppComponent implements AfterViewInit {
   title = 'shop';
+
+  @ViewChild("appTitle")
+  appTitle!: ElementRef<HTMLHeadingElement>;
+
+  ngAfterViewInit(): void {
+    this.appTitle.nativeElement.textContent = "The best shop ever";
+  }
 }