+++ src/app/cart/components/cart-list/cart-list.component.ts
@@ -1,19 +1,18 @@
-import { Component, OnInit } from '@angular/core';
-import { CartItemModel } from '../../models/cart-item.model';
+import { Component } from '@angular/core';
 import { CartService } from '../../services/cart.service';
+import { ProductModel } from 'src/app/products/models/product.model';
 
 @Component({
   selector: 'app-cart-list',
   templateUrl: './cart-list.component.html',
-  styleUrls: ['./cart-list.component.css']
+  styleUrls: ['./cart-list.component.css'],
 })
-export class CartListComponent implements OnInit {
-  cartItems: Array<CartItemModel> = [];
+export class CartListComponent {
 
   constructor(private cartService: CartService) { }
 
-  ngOnInit(): void {
-    this.cartItems = this.cartService.cartItems;
+  get cartItems() {
+    return this.cartService.cartItems;
   }
 
   onClearCart(): void {
@@ -21,6 +20,26 @@ export class CartListComponent implements OnInit {
   }
 
   isCartEmpty(): boolean {
-    return !this.cartItems.length;
+    return !this.cartService.cartItems.length;
+  }
+
+  getTotalCost(): number {
+    return this.cartService.getTotalCost();
+  }
+
+  getTotalQuantity(): number {
+    return this.cartService.getTotalQuantity();
+  }
+
+  onQuantityIncrease(product: ProductModel): void {
+    this.cartService.addToCart(product);
+  }
+
+  onQuantityDecrease(product: ProductModel): void {
+    this.cartService.decreaseCount(product);
+  }
+
+  onDeleteItem(product: ProductModel): void {
+    this.cartService.deleteItem(product);
   }
 }