+++ src/app/products/models/product.model.ts
@@ -1,5 +1,9 @@
 export class ProductModel {
-    
-    constructor(public id: number, public name: string, public price: number) {
+
+    constructor(
+        public id: number,
+        public name: string,
+        public price: number,
+        public isAvailable: boolean) {
     }
 }