+++ src/app/products/services/product.service.ts
@@ -6,9 +6,9 @@ import { ProductModel } from '../models/product.model';
 })
 export class ProductService {
   products: Array<ProductModel> = [
-    new ProductModel(1, "product1", 1),
-    new ProductModel(2, "product2", 2),
-    new ProductModel(3, "product3", 3),
+    new ProductModel(1, "product1", 1, true),
+    new ProductModel(2, "product2", 2, false),
+    new ProductModel(3, "product3", 3, true),
   ]
 
   constructor() { }